+++ lib/background_worker.rb
@@ -24,16 +24,14 @@ module BackgroundWorker
   end
 
   def self.verify_active_connections!
-    Rails.cache.reconnect if defined?(Rails)
-    if defined?(ActiveRecord) && ActiveRecord::VERSION::MAJOR == 3
-      ActiveRecord::Base.verify_active_connections!
+    if defined?(Rails)
+      Rails.cache.reconnect if Rails.cache.respond_to?(:reconnect)
+      Rails.cache.redis.close if Rails.cache.respond_to?(:redis)
     end
   end
 
   def self.release_connections!
-    if defined?(ActiveRecord) && ActiveRecord::VERSION::MAJOR >= 4
-      ActiveRecord::Base.clear_all_connections!
-    end
+    ActiveRecord::Base.clear_all_connections!
   end
 
   def self.after_exception(e)